feat: 고객의 모든 환전 요청을 그룹화하여 조회
1. 고객이 수행한 환전 요청 데이터들의 총 개수 & 해당 고객이 환전을 요청한 총 금액 구하여 출력

+++ src/main/java/com/sparta/currency_user/dto/ExchangeRequestDto.java
@@ -7,15 +7,17 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ExchangeRequestDto {
     @NotBlank(message = "통화 이름을 입력하세요.")
     private final String currenyName;
 
     @NotNull(message = "환전할 금액을 입력해주세요.")
-    private final Long money;
+    private final BigDecimal money;
 
-    public ExchangeRequestDto(String currenyName, Long money){
+    public ExchangeRequestDto(String currenyName, BigDecimal money){
         this.currenyName = currenyName;
         this.money = money;
     }
