+++ src/main/java/com/sparta/currency_user/dto/ExchangeResponseDto.java
@@ -12,13 +12,13 @@ public class ExchangeResponseDto {
     private final Long id;
     private final Long userId;
     private final Long toCurrencyId;
-    private final Long amountInKrw;
+    private final BigDecimal amountInKrw;
     private final BigDecimal amountAfterExchange;
     private final String status;
     private final LocalDateTime createdAt;
     private final LocalDateTime modifiedAt;
 
-    public ExchangeResponseDto(Long id, Long userId, Long toCurrencyId, Long amountInKrw, BigDecimal amountAfterExchange, String status, LocalDateTime createdAt, LocalDateTime modifiedAt) {
+    public ExchangeResponseDto(Long id, Long userId, Long toCurrencyId, BigDecimal amountInKrw, BigDecimal amountAfterExchange, String status, LocalDateTime createdAt, LocalDateTime modifiedAt) {
         this.id = id;
         this.userId = userId;
         this.toCurrencyId = toCurrencyId;
